rm old js hero
