+++ src/pages/Landing/Hero.tsx
@@ -11,10 +11,14 @@ import { Link } from "../../components/Link";
 // @ts-ignore
 import Animate from "animate.css-react";
 
+interface mobile {
+  isMobile: boolean;
+}
+
 const RainbowBackground = styled.div`
   background-image: ${p =>
     `radial-gradient(circle at 100% -80%, ${p.theme.rainbowLight}`});
-  min-height: ${(p: { isMobile: boolean }) => p.isMobile && "100vh"};
+  min-height: ${(p: mobile) => p.isMobile && "calc(100vh - 20px)"};
 `;
 const MainContainer = styled.div`
   max-width: ${p => p.theme.screenSizes.largest};
@@ -22,20 +26,18 @@ const MainContainer = styled.div`
   margin: 0 auto;
   padding: 0 120px 100px;
   @media only screen and (max-width: ${p => p.theme.screenSizes.largest}) {
-    padding: ${(p: { isMobile: boolean }) =>
-      p.isMobile ? "20px" : "0 60px 100px"};
-    min-height: ${(p: { isMobile: boolean }) =>
-      p.isMobile ? "100vh" : undefined};
+    padding: ${(p: mobile) => (p.isMobile ? "20px" : "0 60px 100px")};
+    min-height: ${(p: mobile) => (p.isMobile ? "100vh" : undefined)};
   }
 `;
 const ResponsiveContainer = styled.div`
-  height: ${(p: { isMobile: boolean }) => (p.isMobile ? "100%" : undefined)};
+  height: ${(p: mobile) => (p.isMobile ? "100%" : undefined)};
   > div.is-mobile {
     padding: 0;
     display: flex;
     flex-direction: column;
     justify-content: space-between;
-    min-height: 100vh;
+    min-height: calc(100vh - 20px);
   }
   @media only screen and (min-width: ${p => p.theme.screenSizes.medium}) {
     max-width: 65%;
@@ -54,16 +56,13 @@ const LogoText = styled(Text)`
   font-weight: bold;
 `;
 
-export const Hero = (): JSX.Element => {
+export const Hero = () => {
   const m: boolean = (window as any).mobileCheck();
   return (
     <RainbowBackground isMobile={m}>
       <MainContainer isMobile={m}>
         <ResponsiveContainer isMobile={m}>
-          <div
-            className={`pt100 ${m ? "is-mobile" : undefined}`}
-            id="parent-selector"
-          >
+          <div className={`pt100 ${m ? "is-mobile" : undefined}`}>
             <div>
               <ScrollAnimation animateIn="fadeIn" delay={150} animateOnce>
                 <LogoContainer className={m ? "mb50" : undefined}>
@@ -72,8 +71,8 @@ export const Hero = (): JSX.Element => {
                 </LogoContainer>
                 <Heading
                   level={m ? 1 : 2}
-                  size={m ? "medium" : "large"}
-                  color="blueDark"
+                  size={m ? "meduim" : "large"}
+                  color="brand"
                   className="my20"
                 >
                   Become a validator and help secure the eth2 network.
@@ -93,7 +92,7 @@ export const Hero = (): JSX.Element => {
             >
               <Link
                 to={routesEnum.AcknowledgementPage}
-                className={`${m ? "mb90" : undefined}`}
+                className={`${m ? "mb125" : undefined}`}
               >
                 <Button
                   fullWidth={m}
